Add animation to tile state for non-road tiles
This is a workaround that fixes re-triggering of animations. With the extra
state a class name can be used to limit animations to recently generated tiles.

This is a step in the right direction, as FSM-based animations didn't work as well
as I'd hoped (for road tiles). Eventually all tile animations should use a system
similar to this.